Ajusta a separação da lógica da cartela e do cliente
Close #17
